back-end/requirements.txt
Flask==1.1.2
uWSGI==2.0.19.1
Flask-SQLAlchemy==2.5.1
marshmallow-sqlalchemy==0.26.1
pandas==1.3.4
psycopg2-binary==2.9.1
marshmallow-sqlalchemy==0.26.1
python-dotenv==0.19.1
flask-marshmallow==0.14.0